requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
Authlib==1.3.0
python-dotenv==1.0.0
requests==2.31.0